pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "threat-intel-aggregator"
version = "1.0.0"
description = "AI swarm-powered threat intelligence correlation platform"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # Upper bounds pin against the next major — floors-only invites silent
    # breakage when anthropic / langgraph cut a breaking minor.
    "anthropic>=0.40.0,<2.0",
    "langgraph>=0.2.0,<1.0",
    "langchain-core>=0.3.0,<1.0",
    "pydantic>=2.7.0,<3.0",
    "aiohttp>=3.9.0,<4.0",
    "aiosqlite>=0.20.0,<1.0",
    "fastapi>=0.115.0,<1.0",
    "uvicorn[standard]>=0.30.0,<1.0",
    "structlog>=24.0.0,<26.0",
    "python-dotenv>=1.0.0,<2.0",
    "pyyaml>=6.0.0,<7.0",
    # prometheus_exporter.py lazy-imports at runtime (graceful skip if absent),
    # but the core install ships it so swarm runs can push metrics without an
    # extras install step.
    "prometheus-client>=0.20.0,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "aioresponses>=0.7.6",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "types-PyYAML",
]

[project.scripts]
threat-intel = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E","W","F","I","B","C4","UP","SIM","TCH","RUF"]
# E501  line length — we pick readability over column count
# B008  function call in default arg — Pydantic `Field(...)` triggers it
# RUF012 class-var mutable default — Pydantic `Field(default_factory=list)` triggers it
# TC001/002/003 type-only imports — with `from __future__ import annotations`
#   all annotations are strings anyway, so moving imports into `TYPE_CHECKING`
#   blocks is pure cosmetic churn (7+ files); skip.
ignore = ["E501","B008","RUF012","TC001","TC002","TC003"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short --cov=src --cov-report=term-missing"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = ["pragma: no cover","if TYPE_CHECKING:","raise NotImplementedError"]
